% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_algorithm.r
\name{compute_GAIC}
\alias{compute_GAIC}
\title{Compute GAIC}
\usage{
compute_GAIC(inner_euclidean, trace_S, nu = 2)
}
\arguments{
\item{inner_euclidean}{The inner Euclidean vector.}

\item{trace_S}{The trace of the S matrix.}

\item{nu}{The parameter for penalization.}
}
\value{
The GAIC value.
}
\description{
This function computes the Generalized Akaike Information Criterion (GAIC) used in the RS algorithm.
It calculates GAIC as ||epsilon||^2 + nu * tr(S), where epsilon is the residuals, S is the S matrix, and nu is a parameter.
}
