% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_algorithm.r
\name{compute_GCV}
\alias{compute_GCV}
\title{Compute GCV}
\usage{
compute_GCV(n, inner_euclidean, trace_S)
}
\arguments{
\item{n}{The number of observations.}

\item{inner_euclidean}{The inner Euclidean vector.}

\item{trace_S}{The trace of the S matrix.}
}
\value{
The GCV value.
}
\description{
This function computes the Generalized Cross Validation (GCV) criterion used in the RS algorithm.
It calculates GCV as n * ||epsilon||^2 / (n - tr(S))^2, where epsilon is the residuals and S is the S matrix.
}
