% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_algorithm.r
\name{compute_lambda_gcv}
\alias{compute_lambda_gcv}
\title{Optimise GCV}
\usage{
compute_lambda_gcv(lambda_mu_flag, design_matrix_Z, W, w, z, K, lambda_acc, n)
}
\arguments{
\item{lambda_mu_flag}{A flag indicating whether lambda is for mu or sigma estimation.}

\item{design_matrix_Z}{The B-spline design matrix.}

\item{W}{The weight matrix.}

\item{w}{A vector containing weights.}

\item{z}{A vector of values.}

\item{K}{The penalty matrix.}

\item{lambda_acc}{The tuning parameter for the optimization method.}

\item{n}{The number of observations.}
}
\value{
The optimal value of lambda for GCV.
}
\description{
This function optimizes the GCV criterion to find the optimal value of lambda.
It uses the Brent method for optimization.
}
