% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_algorithm.r
\name{compute_S}
\alias{compute_S}
\title{Compute S Matrix}
\usage{
compute_S(Z, W, lambda, K)
}
\arguments{
\item{Z}{The B-spline design matrix.}

\item{W}{The weight matrix.}

\item{lambda}{The current value of lambda.}

\item{K}{The penalty matrix.}
}
\value{
The S matrix.
}
\description{
This function computes the S matrix used in the RS algorithm.
It calculates S as Z * inv(t(Z) * W * Z + lambda * K) * t(Z) * W.
}
