% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_splines.r
\name{cox_de_boor}
\alias{cox_de_boor}
\title{Cox de Boor Algorithm}
\usage{
cox_de_boor(position, knots, lower, upper)
}
\arguments{
\item{position}{The position at which to evaluate the basis function.}

\item{knots}{The sequence of knots for the splines.}

\item{lower}{The lower index of the basis function.}

\item{upper}{The upper index of the basis function.}
}
\value{
The value of the B-spline basis function at the given position.
}
\description{
This function implements the Cox de Boor algorithm for calculating B-spline basis functions.
It computes the value of the B-spline basis function at a given position within a specific
range defined by lower and upper.
}
