% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_splines.r
\name{create_basis_function_matrix}
\alias{create_basis_function_matrix}
\title{B-Spline Basis Function Matrix}
\usage{
create_basis_function_matrix(
  knot_count,
  lower_bound,
  upper_bound,
  poly_degree,
  vector_x
)
}
\arguments{
\item{knot_count}{The number of knots for the splines.}

\item{lower_bound}{The lower bound of the spline range.}

\item{upper_bound}{The upper bound of the spline range.}

\item{poly_degree}{The degree of the polynomial for the B-splines.}

\item{vector_x}{The input vector of x values.}
}
\value{
The B-spline basis function matrix (design matrix Z).
}
\description{
This function combines the B-spline basis function vectors into a matrix.
Each column of the matrix corresponds to a different basis function.
The resulting matrix is the design matrix Z for the B-splines.
}
