% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_splines.r
\name{create_basis_function_vector}
\alias{create_basis_function_vector}
\title{B-Spline Basis Function Vector}
\usage{
create_basis_function_vector(
  knots_sequence,
  vector_x,
  poly_degree,
  basis_function_knot_position
)
}
\arguments{
\item{knots_sequence}{The sequence of knots for the splines.}

\item{vector_x}{The input vector of x values.}

\item{poly_degree}{The degree of the polynomial for the B-splines.}

\item{basis_function_knot_position}{The position of the basis function within the knot sequence.}
}
\value{
The B-spline basis function vector.
}
\description{
This function computes the B-spline basis function vector for a given set of knots and
a vector of x values.
}
