% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_splines.r
\name{create_diff_matrix}
\alias{create_diff_matrix}
\title{Difference Matrix}
\usage{
create_diff_matrix(order_of_difference_matrix, poly_degree, knot_count)
}
\arguments{
\item{order_of_difference_matrix}{The order of the difference matrix.}

\item{poly_degree}{The degree of the polynomial for the B-splines.}

\item{knot_count}{The number of knots for the splines.}
}
\value{
The difference matrix.
}
\description{
This function creates the difference matrix used in penalized B-splines.
The order of the difference matrix determines the number of differences between
adjacent B-spline basis functions.
}
