% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_splines.r
\name{create_first_order_diff_matrix}
\alias{create_first_order_diff_matrix}
\title{First Order Difference Matrix}
\usage{
create_first_order_diff_matrix(dimensions)
}
\arguments{
\item{dimensions}{The dimensions of the difference matrix.}
}
\value{
The first order difference matrix.
}
\description{
This function creates the first order difference matrix used in penalized B-splines.
It is a square matrix with dimensions (poly_degree + knot_count - 2) x (poly_degree + knot_count - 1).
The matrix contains -1 and 1 along the diagonal and zeros elsewhere.
}
