% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_splines.r
\name{create_knots_for_cox_de_boor}
\alias{create_knots_for_cox_de_boor}
\title{Creating Penalized B-Splines}
\usage{
create_knots_for_cox_de_boor(knot_count, lower_bound, upper_bound, poly_degree)
}
\arguments{
\item{knot_count}{The number of knots (including inner and outer knots).}

\item{lower_bound}{The lower bound of the spline range.}

\item{upper_bound}{The upper bound of the spline range.}

\item{poly_degree}{The degree of the polynomial.}
}
\value{
The sequence of knots for the splines.
}
\description{
This function calculates the knots of the splines using a set of equidistant knots.
The number of inner knots is determined by knot_count, while the number of outer knots
is twice the degree of the polynomial.
}
