% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_splines.r
\name{create_penalty_matrix_K}
\alias{create_penalty_matrix_K}
\title{Penalty Matrix K}
\usage{
create_penalty_matrix_K(order_of_difference_matrix, poly_degree, knot_count)
}
\arguments{
\item{order_of_difference_matrix}{The order of the difference matrix.}

\item{poly_degree}{The degree of the polynomial for the B-splines.}

\item{knot_count}{The number of knots for the splines.}
}
\value{
The penalty matrix K.
}
\description{
This function computes the penalty matrix K used in penalized B-splines.
The penalty matrix is calculated by multiplying the transpose of the difference matrix
with the difference matrix itself.
}
