% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_testing_data.R
\name{generate_polynomial_data}
\alias{generate_polynomial_data}
\title{Generate polynomial data}
\usage{
generate_polynomial_data(n, degree, noise_sd, lower_boundary, upper_boundary)
}
\arguments{
\item{n}{The number of data points to generate.}

\item{degree}{The degree of the polynomial.}

\item{noise_sd}{The standard deviation of the Gaussian noise.}

\item{lower_boundary}{The lower boundary of the x values.}

\item{upper_boundary}{The upper boundary of the x values.}
}
\value{
A data frame with columns 'x' and 'y'.
}
\description{
Generates a dataset with polynomial relationship between x and y, with additive Gaussian noise.
}
\examples{
generate_polynomial_data(100, 3, 0.4, -5, 5)
}
