% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_testing_data.R
\name{generate_polynomial_data_with_changing_sd_linearly}
\alias{generate_polynomial_data_with_changing_sd_linearly}
\title{Generate polynomial data with linearly increasing noise standard deviation}
\usage{
generate_polynomial_data_with_changing_sd_linearly(
  n,
  degree,
  noise_sd_start,
  noise_sd_end,
  lower_boundary,
  upper_boundary
)
}
\arguments{
\item{n}{The number of data points to generate.}

\item{degree}{The degree of the polynomial.}

\item{noise_sd_start}{The starting value of the noise standard deviation.}

\item{noise_sd_end}{The ending value of the noise standard deviation.}

\item{lower_boundary}{The lower boundary of the x values.}

\item{upper_boundary}{The upper boundary of the x values.}
}
\value{
A data frame with columns 'x' and 'y'.
}
\description{
Generates a dataset with polynomial relationship between x and y, where the noise standard deviation
linearly increases from a start value to an end value.
}
\examples{
generate_polynomial_data_with_changing_sd_linearly(100, 3, 0.1, 1, -5, 5)
}
