% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_testing_data.R
\name{generate_polynomial_data_with_changing_sd_oscilating}
\alias{generate_polynomial_data_with_changing_sd_oscilating}
\title{Generate polynomial data with oscillating noise standard deviation}
\usage{
generate_polynomial_data_with_changing_sd_oscilating(
  n,
  degree,
  noise_sd_min,
  noise_sd_max,
  lower_boundary,
  upper_boundary
)
}
\arguments{
\item{n}{The number of data points to generate.}

\item{degree}{The degree of the polynomial.}

\item{noise_sd_min}{The minimum value of the noise standard deviation.}

\item{noise_sd_max}{The maximum value of the noise standard deviation.}

\item{lower_boundary}{The lower boundary of the x values.}

\item{upper_boundary}{The upper boundary of the x values.}
}
\value{
A data frame with columns 'x' and 'y'.
}
\description{
Generates a dataset with polynomial relationship between x and y, where the noise standard deviation
oscillates based on the sine function.
}
\examples{
generate_polynomial_data_with_changing_sd_oscilating(100, 3, 0.1, 0.6, -5, 5)
}
