% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_algorithm.r
\name{global_deviance}
\alias{global_deviance}
\title{Global Deviance}
\usage{
global_deviance(response_vector_y, mu_hat, sigma_hat)
}
\arguments{
\item{response_vector_y}{The response vector of y values.}

\item{mu_hat}{The current estimate of mu_hat.}

\item{sigma_hat}{The current estimate of sigma_hat.}
}
\value{
The twice negative log-likelihood.
}
\description{
This function calculates the twice negative log-likelihood of the model.
It is used to check for convergence within the RS algorithm.
}
