% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_splines.r
\name{indicator}
\alias{indicator}
\title{Indicator Function}
\usage{
indicator(position, lower, upper)
}
\arguments{
\item{position}{The position value to evaluate.}

\item{lower}{The lower bound value.}

\item{upper}{The upper bound value.}
}
\value{
An indicator value (1 or 0) based on the position relative to the lower and upper bounds.
}
\description{
This function returns an indicator value based on the position relative to lower and upper bounds.
It returns 1 if the position is greater than or equal to the lower bound and less than the upper bound, and 0 otherwise.
}
