% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_algorithm.r
\name{init}
\alias{init}
\title{Initialize Starting Values}
\usage{
init(vector_x, response_vector_y)
}
\arguments{
\item{vector_x}{The input vector of x values.}

\item{response_vector_y}{The response vector of y values.}
}
\value{
A list containing the initial values of mu_hat, sigma_hat, beta_hat, gamma_hat, and the design matrix Z.
}
\description{
This function initializes the starting values for the RS algorithm.
It uses a simple linear model to estimate the initial values of mu_hat and sigma_hat.
The function returns a list containing the initial values.
}
