% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_algorithm.r
\name{inner_mu}
\alias{inner_mu}
\title{Inner function for mu estimation}
\usage{
inner_mu(
  response_vector_y,
  mu_hat,
  sigma_hat,
  design_matrix_Z,
  beta_hat,
  lambda_mu,
  K_mu,
  lambda_acc
)
}
\arguments{
\item{response_vector_y}{A vector of response values.}

\item{mu_hat}{The initial estimate of the mean parameter.}

\item{sigma_hat}{The estimate of the variance parameter.}

\item{design_matrix_Z}{The design matrix of B-spline basis functions.}

\item{beta_hat}{The current estimate of the B-spline coefficients.}

\item{lambda_mu}{The tuning parameter (lambda) for updating the mean parameter.}

\item{K_mu}{The penalty matrix for the mean parameter.}

\item{lambda_acc}{The tuning parameter for the optimization method.}
}
\value{
A list containing the updated mean parameter (mu_hat), updated B-spline coefficients (beta_hat),
and the optimal lambda_mu value.
}
\description{
This function performs the inner loop for estimating the mean parameter (mu_hat) in the Local Maximum Likelihood (LML) regression model.
It updates the beta_hat coefficients using Penalized Weighted Least Squares (PWLS) and iteratively updates mu_hat until convergence is reached.
}
