% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_algorithm.r
\name{inner_sigma}
\alias{inner_sigma}
\title{Inner function for sigma estimation}
\usage{
inner_sigma(
  mu_hat,
  sigma_hat,
  design_matrix_Z,
  response_vector_y,
  gamma_hat,
  lambda_sigma,
  K_sigma
)
}
\arguments{
\item{mu_hat}{The current estimate of the mean parameter.}

\item{sigma_hat}{The initial estimate of the variance parameter.}

\item{design_matrix_Z}{The design matrix of B-spline basis functions.}

\item{response_vector_y}{A vector of response values.}

\item{gamma_hat}{The current estimate of the gamma coefficients for variance.}

\item{lambda_sigma}{The tuning parameter (lambda) for updating the variance parameter.}

\item{K_sigma}{The penalty matrix for the variance parameter.}
}
\value{
A list containing the updated variance parameter (sigma_hat), updated gamma coefficients (gamma_hat),
and the optimal lambda_sigma value.
}
\description{
This function performs the inner loop for estimating the variance parameter (sigma_hat) in the Local Maximum Likelihood (LML) regression model.
It updates the gamma_hat coefficients using Penalized Weighted Least Squares (PWLS) and iteratively updates sigma_hat until convergence is reached.
}
