% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_algorithm.r
\name{optimise_GAIC}
\alias{optimise_GAIC}
\title{Optimise GAIC}
\usage{
optimise_GAIC(lambda_mu_flag, Z, W, lambda, K, epsilon, beta_hat, nu = 2)
}
\arguments{
\item{lambda_mu_flag}{A flag indicating whether lambda is for mu or sigma estimation.}

\item{Z}{The B-spline design matrix.}

\item{W}{The weight matrix.}

\item{lambda}{The current value of lambda.}

\item{K}{The penalty matrix.}

\item{epsilon}{The epsilon vector.}

\item{beta_hat}{The current estimate of beta_hat or gamma_hat.}

\item{nu}{The parameter for penalization.}
}
\value{
The optimal value of lambda for GAIC.
}
\description{
This function optimizes the GAIC criterion to find the optimal value of lambda.
It uses the Brent method for optimization.
}
