% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_algorithm.r
\name{optimise_GCV}
\alias{optimise_GCV}
\title{Optimise GCV}
\usage{
optimise_GCV(lambda_mu_flag, Z, W, w, lambda, K, epsilon, beta_hat, n)
}
\arguments{
\item{lambda_mu_flag}{A flag indicating whether lambda is for mu or sigma estimation.}

\item{Z}{The B-spline design matrix.}

\item{W}{The weight matrix.}

\item{w}{The weights.}

\item{lambda}{The current value of lambda.}

\item{K}{The penalty matrix.}

\item{epsilon}{The epsilon vector.}

\item{beta_hat}{The current estimate of beta_hat or gamma_hat.}

\item{n}{The number of observations.}
}
\value{
The optimal value of lambda for GCV.
}
\description{
This function optimizes the GCV criterion to find the optimal value of lambda.
It uses the Brent method for optimization.
}
