% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_algorithm.r
\name{outer}
\alias{outer}
\title{Successively updates the mean and variance parameters}
\usage{
outer(
  mu_hat,
  sigma_hat,
  design_matrix_Z,
  vector_x,
  response_vector_y,
  beta_hat,
  gamma_hat,
  lambda_mu,
  lambda_sigma,
  max_iterations,
  K_mu,
  K_sigma,
  knot_count,
  poly_degree,
  opt_method,
  lambda_acc
)
}
\arguments{
\item{mu_hat}{The initial estimate of the mean parameter.}

\item{sigma_hat}{The initial estimate of the variance parameter.}

\item{design_matrix_Z}{The design matrix of B-spline basis functions.}

\item{vector_x}{A vector of predictor values.}

\item{response_vector_y}{A vector of response values.}

\item{beta_hat}{The initial estimate of the B-spline coefficients.}

\item{gamma_hat}{The initial estimate of the gamma coefficients for variance.}

\item{lambda_mu}{The tuning parameter (lambda) for updating the mean parameter.}

\item{lambda_sigma}{The tuning parameter (lambda) for updating the variance parameter.}

\item{max_iterations}{The maximum number of iterations for the algorithm.}

\item{K_mu}{The penalty matrix for the mean parameter.}

\item{K_sigma}{The penalty matrix for the variance parameter.}

\item{knot_count}{The number of knots in the B-spline basis.}

\item{poly_degree}{The degree of the B-spline polynomials.}

\item{opt_method}{The optimization method to use for tuning parameter selection ('GCV' or 'GAIC').}

\item{lambda_acc}{The tuning parameter for the optimization method.}
}
\value{
A list containing the final estimates of the mean (mu_hat) and variance (sigma_hat),
along with the estimates of the B-spline coefficients (beta_hat) and gamma coefficients for variance (gamma_hat).
}
\description{
This function performs the outer loop of the Local Maximum Likelihood (LML) regression model,
successively updating the mean (mu_hat) and variance (sigma_hat) parameters until convergence is reached.
}
