% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_functions.r
\name{plot_lmls_bspline}
\alias{plot_lmls_bspline}
\title{Plotting function for lmls_bspline results using ggplot2}
\usage{
plot_lmls_bspline(result, data)
}
\arguments{
\item{result}{An object of class "lmls_bspline" containing the result of the lmls_bspline function.}

\item{data}{The original data used in the lmls_bspline function.}
}
\value{
A ggplot2 object.
}
\description{
This function plots the estimated mean function along with confidence intervals
for the lmls_bspline results using ggplot2.
}
\examples{
data <- generate_polynomial_data(100, 3, 0.4, -5, 5)
result <- lmls_bspline(data$x, data$y)
plot_lmls_bspline(result, data)
}
