% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_splines.r
\name{PLS_b_Splines}
\alias{PLS_b_Splines}
\title{Penalized Least Squares B-Splines}
\usage{
PLS_b_Splines(
  knot_count,
  lower_bound,
  upper_bound,
  poly_degree,
  vector_x,
  response_vector_y,
  hyperparameter_lambda
)
}
\arguments{
\item{knot_count}{The number of knots for the splines.}

\item{lower_bound}{The lower bound of the spline range.}

\item{upper_bound}{The upper bound of the spline range.}

\item{poly_degree}{The degree of the polynomial for the B-splines.}

\item{vector_x}{The input vector of x values.}

\item{response_vector_y}{The response vector of y values.}

\item{hyperparameter_lambda}{The hyperparameter for penalization.}
}
\value{
A list containing the B-spline design matrix Z and the coefficients gamma_hat.
}
\description{
This function performs the penalized least squares regression using B-splines.
It computes the B-spline design matrix Z, the penalty matrix K, and the coefficients gamma_hat.
The hyperparameter lambda determines the amount of penalization.
}
