% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_functions.r
\name{summary_lmls_bspline}
\alias{summary_lmls_bspline}
\title{Summary function for lmls_bspline results}
\usage{
summary_lmls_bspline(lmls_object, digits = 3)
}
\arguments{
\item{lmls_object}{An object of class "pSplineLocationScale" obtained from lmls_bspline function.}

\item{digits}{Number of digits to round numerical values in the summary.}
}
\value{
A summary of the lmls_bspline object.
}
\description{
This function provides a summary of the results obtained from the lmls_bspline function.
}
\examples{
data <- generate_polynomial_data(100, 3, 0.4, -5, 5)
result <- lmls_bspline(data$x, data$y)
summary_lmls_bspline(result)
}
